+++ lib/enum/CurlCode.ts
@@ -165,6 +165,7 @@ export enum CurlCode {
   CURLE_AUTH_ERROR /* 94 - an authentication function returned an error */,
   CURLE_HTTP3 /* 95 - An HTTP/3 layer problem */,
   CURLE_QUIC_CONNECT_ERROR /* 96 - QUIC connection error */,
+  CURLE_PROXY /* 97 - proxy handshake error */,
   CURLE_LAST,
   /*  compatibility with older names */
   CURLE_FTP_WEIRD_SERVER_REPLY = CURLE_WEIRD_SERVER_REPLY,