+++ CHANGELOG.md
@@ -54,6 +54,7 @@ As `curly` is marked as experimental, this allows us to do a breaking change in 
   These new features related to streams are only reliable when using a libcurl version >= 7.69.1.
 - Support libcurl info `CURLINFO_CERTINFO`. Can be retrieved using `getInfo("CERTINFO")`. Thanks to [@Sergey-Mityukov](https://github.com/Sergey-Mityukov) for most of the work on this.
 - Support libcurl info `CURLINFO_EFFECTIVE_METHOD`. Requires libcurl >= 7.72.0.
+- Support libcurl info `CURLINFO_PROXY_ERROR`. Use `CurlPx` for constants. Requires libcurl >= 7.73.0.
 
 ### Changed
 - `curly` now has 100% code coverage.